worker: Split into separate files.
This makes each worker faster to start up.
