+++ docs/subsystems/queuing.md
@@ -35,14 +35,12 @@ custom integration defined in `zerver/lib/queue.py`.
 
 To add a new queue processor:
 
-- Define the processor in `zerver/worker/queue_processors.py` using
-  the `@assign_queue` decorator; it's pretty easy to get the template
-  for an existing similar queue processor. This suffices to test your
-  queue worker in the Zulip development environment
-  (`tools/run-dev` will automatically restart the queue processors
-  and start running your new queue processor code). You can also run
-  a single queue processor manually using e.g.
-  `./manage.py process_queue --queue=user_activity`.
+- Define the processor in `zerver/worker/` using the `@assign_queue` decorator;
+  it's pretty easy to get the template for an existing similar queue
+  processor. This suffices to test your queue worker in the Zulip development
+  environment (`tools/run-dev` will automatically restart the queue processors
+  and start running your new queue processor code). You can also run a single
+  queue processor manually using e.g. `./manage.py process_queue --queue=user_activity`.
 
 - So that supervisord will know to run the queue processor in
   production, you will need to add to the `queues` variable in
